Open stream after listeners are added
Currently `error` listener is added after the connection is open.
So if there is a synchronous error during connection open (for example
exception) then we don't receive this error in the error listener.

We found this after noticing that if WebSocket connection to Twilio is
blocked by CSP then on iOS Safari it's not being reported as error.

On Chrome this does not reproduce as it will not throw an error in this
case, but instead call WebSocket error listener.